Rename AnalyticsService file
