+++ examples/analytics/simple.php
@@ -1,6 +1,6 @@
 <?php
 require_once '../../src/apiClient.php';
-require_once '../../src/Contrib/apiAnalyticsService.php';
+require_once '../../src/Contrib/AnalyticsService.php';
 session_start();
 
 $client = new apiClient();
